reference compilation refactor
